contractG/requirements.txt
# 基础UI框架
PyQt5>=5.15.9        # 桌面应用UI框架
PyQt5-sip>=12.9.0    # PyQt5的核心依赖

# 数据处理
pandas>=2.0.0        # Excel文件读写和数据处理
openpyxl>=3.1.2      # Excel文件高级操作和样式设置

# 配置管理
configparser>=6.0.0  # 应用程序配置管理

# 中文支持
pypinyin>=0.49.0     # 中文拼音处理

# PDF处理
PyPDF2>=3.0.0        # PDF文件读写和处理
reportlab>=4.0.0     # PDF生成和编辑
Pillow>=10.0.0        # 图像处理
PyMuPDF>=1.23.0      # PDF转图片处理（可选，用于生成图片式PDF）

# Windows系统支持
pywin32>=306         # Windows COM接口，用于Excel转PDF

# 其他依赖
pyinstaller>=6.0.0

